MAINT: moved getstate methods to BaseModel
