+++ force_bdss/core/kpi_specification.py
@@ -1,6 +1,6 @@
 from traits.api import Enum, HasStrictTraits, Float, Bool
 
-from force_bdss.io.workflow_writer import pop_dunder_recursive
+from force_bdss.core.base_model import pop_dunder_recursive
 from force_bdss.local_traits import Identifier
 from .verifier import VerifierError
