+++ force_bdss/core/output_slot_info.py
@@ -1,7 +1,7 @@
 from traits.api import HasStrictTraits
 
 from force_bdss.core.verifier import VerifierError
-from force_bdss.io.workflow_writer import pop_dunder_recursive
+from force_bdss.core.base_model import pop_dunder_recursive
 from force_bdss.local_traits import Identifier
