+++ force_bdss/core/slot.py
@@ -1,6 +1,6 @@
 from traits.api import HasStrictTraits, Unicode
 from ..local_traits import CUBAType
-from force_bdss.io.workflow_writer import pop_dunder_recursive
+from force_bdss.core.base_model import pop_dunder_recursive
 
 
 class Slot(HasStrictTraits):
